restructured reducer
