+++ my-app/package.json
@@ -9,6 +9,7 @@
       "axios": "^0.21.1",
       "faker": "^5.5.3",
       "history": "^5.0.0",
+      "jest": "^26.6.0",
       "mirage": "0.0.7",
       "miragejs": "^0.1.41",
       "react": "^17.0.2",
