+++ my-app/src/contexts/ProductContext.js
@@ -1,5 +1,5 @@
 import { createContext, useContext, useReducer } from "react";
-
+import { reducerFunction } from "../reducers/productReducer";
 export const ProductsContext = createContext();
 
 export function ProductsProvider({ children }) {
@@ -23,56 +23,4 @@ export function ProductsProvider({ children }) {
 
 export function useProducts() {
   return useContext(ProductsContext);
-}
-
-const reducerFunction = (acc, action) => {
-  switch (action.type) {
-    case "SET PRODUCTS":
-      return {
-        ...acc,
-        products: action.payload,
-      };
-    case "SET CART":
-      return { ...acc, cart: action.payload };
-
-    case "SET CART AFTER MOVE":
-      return { ...acc };
-    case "SET WISHLIST":
-      return { ...acc, wishlist: action.payload };
-
-    case "SORT":
-      return { ...acc, sortBy: action.payload };
-    case "PRICE_RANGE":
-      return { ...acc, price_range: action.payload };
-    case "RESET":
-      return { ...acc, sortBy: null, price_range: 0 };
-    case "IN STOCK":
-      return {
-        ...acc,
-        products: { ...acc.products },
-        stock: action.payload,
-      };
-
-    default:
-      break;
-  }
-};
-
-export function getSortedData(productList, sortBy) {
-  if (sortBy && sortBy === "PRICE_HIGH_TO_LOW") {
-    return [...productList].sort((a, b) => b["price"] - a["price"]);
-  }
-  if (sortBy && sortBy === "PRICE_LOW_TO_HIGH") {
-    return [...productList].sort((a, b) => a["price"] - b["price"]);
-  }
-  return productList;
-}
-
-export function getPriceRangeData(productList, price_range) {
-  if (parseInt(price_range, 10) > 0) {
-    return productList.filter(
-      (item) => parseInt(item.price, 10) < parseInt(price_range, 10)
-    );
-  }
-  return productList;
 }
